requirements.txt
asgiref==3.6.0
coverage==7.2.4
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
